Rename handlers to handler
